+++ packages/orq-rc/src/orq_ai_sdk/models/createannotationqueueop.py
@@ -104,7 +104,7 @@ class CreateAnnotationQueueResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-08-24T16:03:44.631Z")
+    updated: Optional[datetime] = parse_datetime("2026-08-25T03:48:43.842Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")