+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -252,9 +252,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-08-24T16:03:46.031Z"
+    created: Optional[str] = "2026-08-25T03:48:45.450Z"
 
-    updated: Optional[str] = "2026-08-24T16:03:46.031Z"
+    updated: Optional[str] = "2026-08-25T03:48:45.450Z"
 
     updated_by_id: OptionalNullable[str] = UNSET