+++ packages/orq-rc/RELEASES.md
@@ -10429,3 +10429,13 @@ Based on:
 - [python v4.14.0-rc.49] packages/orq-rc
 ### Releases
 - [PyPI v4.14.0-rc.49] https://pypi.org/project/orq-ai-sdk/4.14.0-rc.49 - packages/orq-rc
+
+## 2026-08-25 03:51:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.795.3 (2.932.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.14.0-rc.50] packages/orq-rc
+### Releases
+- [PyPI v4.14.0-rc.50] https://pypi.org/project/orq-ai-sdk/4.14.0-rc.50 - packages/orq-rc